mudado a apresentação da hora e a view dos empregados

+++ app/models/book.rb
@@ -5,7 +5,7 @@ class Book < ActiveRecord::Base
   belongs_to :student
 
   def formated_hour
-    self.hour.to_formatted_s(:short)
+    self.hour.to_formatted_s(:long)
   end
 
   def self.not_booked
@@ -13,6 +13,7 @@ class Book < ActiveRecord::Base
   end
 
   private
+  #VERIFICAR PQ NÃO TÁ APARECENDO UM ERRO LÁ NA VIEW
   def check_date
     if (self.hour < Time.now)
       self.errors.add(:notice, "registre apenas dias para frente")
